+++ Common/changelog.md
@@ -1,5 +1,11 @@
 ## Version 1.9.0
-- Make all plushies wearable on the head
+- Made all plushies wearable on the head
+- New plushies:
+  - Robin
+  - Otter
+- Made player plushies send a message under the name of "Player Plushie" instead of plushie
+- Made dropped duck plushie item smaller
+
 
 ## Version 1.8.0
 - Rephrased "Early Supporter" tooltip to "Community Pillar"